test: run tests of all go.mods (#4468)
* test: run tests of all go.mods


* Remove +build directive which is deprecated, remove tparse hadnling in Makefile.

* Add sdk script, remove LINT_DIFF.

* Fix linting issues that popped up.

* Add script to run tests for all modules.

* chore: add executible permissions to script, rename to avoid shadowing dir builtin

* chore: use python3 instead of python

* chore: add venv to gitignore

* Disallow env vars not being set, clean up minor issues.

---------

+++ e2e/tests/upgrades/genesis_test.go
@@ -1,3 +1,5 @@
+//go:build !test_e2e
+
 package upgrades
 
 import (
@@ -194,7 +196,7 @@ func (s *GenesisTestSuite) HaltChainAndExportGenesis(ctx context.Context, chain 
 	err = chain.StopAllNodes(ctx)
 	s.Require().NoError(err, "error stopping node(s)")
 
-	state, err := chain.ExportState(ctx, int64(haltHeight))
+	state, err := chain.ExportState(ctx, haltHeight)
 	s.Require().NoError(err)
 
 	appTomlOverrides := make(test.Toml)
